# docs/swagger.yaml
basePath: /api/v1
definitions:
  productsmodel.Product:
    properties:
      currentstock:
        description: '`json:"currentstock" bson:"currentstock"`'
        type: number
      id:
        type: string
      item:
        description: '`json:"item" bson:"item"`'
        type: string
      standard:
        description: '`json:"standard" bson:"standard"`'
        type: string
      unitprice:
        description: '`json:"unitprice" bson:"unitprice"`'
        type: number
    type: object
  usersmodel.LoginUser:
    properties:
      email:
        description: '`json:"email" bson:"email"`'
        type: string
      password:
        description: '`json:"password" bson:"password"`'
        type: string
    type: object
  usersmodel.SignupUser:
    properties:
      email:
        description: '`json:"email" bson:"email"`'
        type: string
      idnumber:
        description: '`json:"idnumber" bson:"idnumber"`'
        type: integer
      password:
        description: '`json:"password" bson:"password"`'
        type: string
      phonenumber:
        description: '`json:"phonenumber" bson:"phonenumber"`'
        type: string
      username:
        description: '`json:"username" bson:"username"`'
        type: string
    type: object
  usersmodel.SuccessMessage:
    properties:
      message:
        description: '`json:"message" bson:"message"'
        type: string
    type: object
  usersmodel.User:
    properties:
      email:
        description: '`json:"email" bson:"email"`'
        type: string
      id:
        type: string
      idnumber:
        description: '`json:"idnumber" bson:"idnumber"`'
        type: integer
      isactive:
        description: '`json:"isactive" bson:"isactive"`'
        type: boolean
      isadmin:
        description: '`json:"isadmin" bson:"isadmin"`'
        type: boolean
      isblacklisted:
        description: '`json:"isblacklisted" bson:"isblacklisted"`'
        type: boolean
      isvalid:
        description: '`json:"isvalid" bson:"isvalid"`'
        type: boolean
      password:
        description: '`json:"password" bson:"password"`'
        type: string
      phonenumber:
        description: '`json:"phonenumber" bson:"phonenumber"`'
        type: string
      userclass:
        description: '`json:"userclass" bson:"userclass"`'
        type: string
      username:
        description: '`json:"username" bson:"username"`'
        type: string
    type: object
info:
  contact:
    email: email@email.com
    name: API Support
  description: This is a service for connecting farmers to customers
  license:
    name: MIT
    url: https://github.com/maxwellgithinji/farmsale_backend/blob/develop/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Farmsale API
  version: "1.0"
paths:
  /admin/profile/activate/{id}:
    put:
      consumes:
      - application/json
      description: ActivateDeactivateAccount is a safer option than deleting accounts which have interracted with the application
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usersmodel.SuccessMessage'
      security:
      - ApiKeyAuth: []
      summary: Admin activates or deactivates the account
      tags:
      - admin
  /admin/profile/delete/{id}:
    put:
      consumes:
      - application/json
      description: this completely removes user account data
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usersmodel.SuccessMessage'
      security:
      - ApiKeyAuth: []
      summary: DeleteUser is for admin purposes in case an account was wrongfuly created
      tags:
      - admin
  /auth/products:
    get:
      description: Get details of all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/productsmodel.Product'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get details of all products
      tags:
      - auth User
  /login:
    post:
      consumes:
      - application/json
      description: Login an existing user with their credentials
      parameters:
      - description: login user
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/usersmodel.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usersmodel.User'
      summary: Logs in an existing User
      tags:
      - auth
  /profile/{id}:
    put:
      consumes:
      - application/json
      description: Editing is only accessible to the owners of the credentials
      parameters:
      - description: edit profile
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/usersmodel.SignupUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usersmodel.SuccessMessage'
      security:
      - ApiKeyAuth: []
      summary: A user is able to edit their account details
      tags:
      - user
  /profile/deactivate/{id}:
    put:
      consumes:
      - application/json
      description: DeactivateAccount is a safer option than deleting accounts which have interracted with the application
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usersmodel.SuccessMessage'
      security:
      - ApiKeyAuth: []
      summary: Owner of the account activates or deactivates the account
      tags:
      - user
  /signup:
    post:
      consumes:
      - application/json
      description: Signs Up a user with new credentials
      parameters:
      - description: signup user
        in: body
        name: signup
        required: true
        schema:
          $ref: '#/definitions/usersmodel.SignupUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/usersmodel.User'
      summary: Signs up a new user
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"